- moved some files - removed almost all imports from app
