+++ frontend/src/components/layout/ElectionLayout.tsx
