+++ frontend/src/components/page/FatalErrorPage.tsx
