+++ frontend/src/features/apportionment/components/full_seats/ApportionmentFullSeatsPage.test.tsx
@@ -1,8 +1,7 @@
 import { render as rtlRender } from "@testing-library/react";
 import { describe, expect, test, vi } from "vitest";
 
-// eslint-disable-next-line import/no-restricted-paths -- #1283
-import { routes } from "@/app/routes";
+import { ErrorBoundary } from "@/components/error/ErrorBoundary";
 import { ElectionProvider } from "@/hooks/election/ElectionProvider";
 import { getElectionMockData } from "@/testing/api-mocks/ElectionMockData";
 import { Providers } from "@/testing/Providers";
@@ -12,6 +11,7 @@ import { ElectionApportionmentResponse, ErrorResponse } from "@/types/generated/
 
 import * as lt19Seats from "../../testing/lt-19-seats";
 import * as lt19SeatsAndP9AndP10 from "../../testing/lt-19-seats-and-p9-and-p10";
+import { apportionmentRoutes } from "../../routes";
 import { ApportionmentProvider } from "../ApportionmentProvider";
 import { ApportionmentFullSeatsPage } from "./ApportionmentFullSeatsPage";
 
@@ -205,8 +205,13 @@ describe("ApportionmentFullSeatsPage", () => {
       vi.spyOn(console, "error").mockImplementation(() => {
         /* do nothing */
       });
-      const router = setupTestRouter(routes);
-
+      const router = setupTestRouter([
+        {
+          element: null,
+          errorElement: <ErrorBoundary />,
+          children: apportionmentRoutes,
+        },
+      ]);
       overrideOnce("get", "/api/elections/1", 200, getElectionMockData(lt19Seats.election));
       overrideOnce("post", "/api/elections/1/apportionment", 500, {
         error: "Internal Server Error",
