+++ frontend/src/features/data_entry/routes.tsx
@@ -1,6 +1,5 @@
 import { RouteObject } from "react-router";
 
-import { DataEntryChoicePage } from "../data_entry_choice/components/DataEntryChoicePage";
 import { CandidatesVotesPage } from "./components/candidates_votes/CandidatesVotesPage";
 import { CheckAndSavePage } from "./components/check_and_save/CheckAndSavePage";
 import { DataEntryLayout } from "./components/DataEntryLayout";
@@ -8,32 +7,23 @@ import { DifferencesPage } from "./components/differences/DifferencesPage";
 import { RecountedPage } from "./components/recounted/RecountedPage";
 import { VotersAndVotesPage } from "./components/voters_and_votes/VotersAndVotesPage";
 
-//TODO: DataEntryChoice page is odd one out
-
 export const dataEntryRoutes: RouteObject[] = [
   {
-    path: "data-entry",
-    element: null,
+    path: ":pollingStationId/:entryNumber",
+    element: <DataEntryLayout />,
     children: [
-      { index: true, element: <DataEntryChoicePage /> },
+      { index: true, element: null },
+      { path: "recounted", element: <RecountedPage /> },
       {
-        path: ":pollingStationId/:entryNumber",
-        element: <DataEntryLayout />,
-        children: [
-          { index: true, element: null },
-          { path: "recounted", element: <RecountedPage /> },
-          {
-            path: "voters-and-votes",
-            element: <VotersAndVotesPage />,
-          },
-          { path: "differences", element: <DifferencesPage /> },
-          {
-            path: "list/:listNumber",
-            element: <CandidatesVotesPage />,
-          },
-          { path: "save", element: <CheckAndSavePage /> },
-        ],
+        path: "voters-and-votes",
+        element: <VotersAndVotesPage />,
       },
+      { path: "differences", element: <DifferencesPage /> },
+      {
+        path: "list/:listNumber",
+        element: <CandidatesVotesPage />,
+      },
+      { path: "save", element: <CheckAndSavePage /> },
     ],
   },
 ];
