+++ frontend/src/features/data_entry_choice/routes.tsx
@@ -1,0 +1,5 @@
+import { RouteObject } from "react-router";
+
+import { DataEntryChoicePage } from "./components/DataEntryChoicePage";
+
+export const dataEntryChoiceRoutes: RouteObject[] = [{ index: true, element: <DataEntryChoicePage /> }];
