+++ frontend/src/features/dev/components/DevHomePage.tsx
@@ -1,14 +1,13 @@
 import { Link } from "react-router";
 
 import { useApiState } from "@/api/useApiState";
+import { MockTest } from "@/components/mock/MockTest";
 import { PageTitle } from "@/components/page_title/PageTitle";
 import { AppLayout } from "@/components/ui/AppLayout/AppLayout";
 import { ElectionListProvider } from "@/hooks/election/ElectionListProvider";
 import { useElectionList } from "@/hooks/election/useElectionList";
 import { useUserRole } from "@/hooks/user/useUserRole";
 import { t } from "@/lib/i18n";
-
-import { MockTest } from "./MockTest";
 
 function TypistLinks() {
   const { electionList } = useElectionList();
