+++ frontend/src/features/dev/routes.tsx
@@ -1,0 +1,5 @@
+import { RouteObject } from "react-router";
+
+import { DevHomePage } from "./components/DevHomePage";
+
+export const devRoutes: RouteObject[] = [{ path: "dev", element: <DevHomePage /> }];
