+++ frontend/src/features/logs/routes.tsx
@@ -1,0 +1,5 @@
+import { RouteObject } from "react-router";
+
+import { LogsHomePage } from "./components/LogsHomePage";
+
+export const logsRoutes: RouteObject[] = [{ path: "logs", element: <LogsHomePage /> }];
