+++ frontend/src/features/users/components/create/UserCreate.test.tsx
@@ -2,15 +2,21 @@ import { render as rtlRender } from "@testing-library/react";
 import { userEvent } from "@testing-library/user-event";
 import { beforeEach, describe, expect, test } from "vitest";
 
-// eslint-disable-next-line import/no-restricted-paths -- #1283
-import { routes } from "@/app/routes";
 import { ElectionListRequestHandler, UserListRequestHandler } from "@/testing/api-mocks/RequestHandlers";
 import { Providers } from "@/testing/Providers";
 import { overrideOnce, server } from "@/testing/server";
 import { screen, setupTestRouter, waitFor, within } from "@/testing/test-utils";
 
+import { usersRoutes } from "../../routes";
+
 function renderWithRouter() {
-  const router = setupTestRouter(routes);
+  const router = setupTestRouter([
+    {
+      path: "/",
+      element: null,
+      children: usersRoutes,
+    },
+  ]);
   rtlRender(<Providers router={router} />);
   return router;
 }
