+++ frontend/src/features/workstations/routes.tsx
@@ -1,0 +1,10 @@
+import { RouteObject } from "react-router";
+
+import { WorkstationsHomePage } from "./components/WorkstationsHomePage";
+
+export const workstationsRoutes: RouteObject[] = [
+  {
+    path: "workstations",
+    element: <WorkstationsHomePage />,
+  },
+];
